tools/python/youtube_info_updater/requirements.txt
mysql-connector-python
python-dotenv
requests